pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "frontier-feed"
version = "0.1.0"
description = "Personal AI-frontier intelligence pipeline → Telegram, reactive, routine-driven"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pyyaml>=6",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "beautifulsoup4>=4.12",
    "groq>=0.11",
    "google-generativeai>=0.8",
    "tavily-python>=0.5",
    "pydantic>=2.8",
]

[project.scripts]
feed = "feed.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["feed*"]
